secp256k1-program: Fix doc tests (#245)
* secp256k1-program: Fix doc tests

#### Problem

The doctests for secp256k1-program fail when run on the crate
individually because of a missing feature declaration on
`solana-keccak-hasher`.

#### Summary of changes

Add the feature. At the same time, run the doctests with `cargo-hack` to
catch these in the future.

The sdk-wasm-js package fails to run doctests because it doesn't have an
rlib target, so add that too.

* Use nightly and install hack for doc test job