program-entrypoint: Add a BumpAllocator constructor (#284)
* program-entrypoint: Add a BumpAllocator constructor

#### Problem

The current version of program-entrypoint doesn't work because the
allocator can't be instantiated in the macros. The start and len fields
are private, and so can't be written to in the macro, which exists in
the scope of the program.

#### Summary of changes

Add a new constructor that can be used by the macro.

* Make function const, call from unsafe block

* Rename