scripts: Allow for patching individual crates (#370)
#### Problem

Now that the crates in the SDK are broken up and don't depend on fixed
versions of each other, it's possible to patch individual crates into an
Agave build, but there's no documentation or scripts to help with that.

#### Summary of changes

Update the patching script to take in an optional path, to patch just
one crate at a time. By default, the script patches everything. It can
be run more than once. Also, document how to use it for Agave devs.