sha256-hasher: Make it no-std compatible (#384)
* sha256-hasher: Make it no-std compatible

#### Problem

The sha256-hasher doesn't compile in real no-std environments.

#### Summary of changes

This was an easy fix, simply don't pull in default features for the sha2
crate. This means we disable OIDs for hashes, and the ability to
finalize hashes into Box'ed values. Since we don't use that, it should
be safe to disable.

Also, the crate was a no-op to build without the sha2 feature, so I
changed the no-std checking strategy to *exclude* features rather than
include certain features, and moved it to `cargo hack` so that each
crate and feature is tested individually.

* Install cargo-hack in CI