pubkey: Make no-std compatible by enabling atomic with std (#388)
* pubkey: Make no-std compatible by enabling atomic with std

#### Problem

The pubkey crate doesn't compile for no-std environments because the
atomic feature is always enabled on address.

This is an alternative to #302.

#### Summary of changes

Add the "atomic" feature to the "std" feature in pubkey, and enable
"std" as appropriate for certain tests.

We could do the same with "decode", but since it isn't required, I'm
inclined to make a minimal change. People can switch to address for more
control.

* instruction: Add std feature for doc build