Remove frozen-abi-macro dependency on solana-logger (#400)
There are several log!() lines as well as the solana_logger::setup() in
frozen-abi-macro. Adjust the log lines to println!() and eprintln!() to
remove the dependency; these output lines are only in frozen-abi tests.
Update frozen-abi to use prints as well since items in that crate are
called by frozen-abi-macro